a2a-backend/requirements.txt
langchain-core
langchain
langchain-openai
langgraph
python-dotenv==1.0.0
httpx==0.27.0
aiohttp==3.9.1
pydantic
typing-extensions
fastapi
uvicorn==0.24.0
websockets==11.0.0
